// express-api/generated/kysely.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export type Numeric = ColumnType<string, number | string>;

export type Timestamp = ColumnType<Date, Date | string>;
// ...
export interface Offers {
  ahref: string | null;
  brand: string | null;
  brand_key: string | null;
  context: string;
  created_at: Generated<Timestamp>;
  currency: string | null;
  dealer_key: string;
  description: string | null;
  difference_180_days_mean: Numeric | null;
  difference_180_days_mean_percentage: Numeric | null;
  difference_30_days_mean: Numeric | null;
  difference_30_days_mean_percentage: Numeric | null;
  difference_365_days_mean: Numeric | null;
  difference_365_days_mean_percentage: Numeric | null;
  difference_7_days_mean: Numeric | null;
  difference_7_days_mean_percentage: Numeric | null;
  difference_90_days_mean: Numeric | null;
  difference_90_days_mean_percentage: Numeric | null;
  href: string;
  image: string | null;
  is_partner: Generated<boolean | null>;
  is_promotion_restricted: boolean | null;
  item_condition: string | null;
  market: string;
  mongo_id: string | null;
  mpn: string | null;
  mpn_stock: string | null;
  pre_price: Numeric | null;
  price: Numeric | null;
  price_unit: string | null;
  prices_migrated_at: Timestamp | null;
  product_id: string;
  provenance: string;
  provenance_id: string;
  quantity_amount: Numeric | null;
  quantity_standard_amount: Numeric | null;
  quantity_unit: string | null;
  scrape_batch_id: string | null;
  short_description: string | null;
  subtitle: string | null;
  title: string;
  upc: string | null;
  updated_at: Generated<Timestamp>;
  uri: string;
  valid_from: Timestamp;
  valid_through: Timestamp;
  value_amount: Numeric | null;
  value_standard_amount: Numeric | null;
  value_unit: string | null;
  vendor_key: string | null;
}
// ...
export interface DB {
// ...
  offers: Offers;
// ...
}